requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
python-dotenv==1.0.0
openpyxl==3.1.2
reportlab==4.0.7